+++ DataStructure/binary_indexed_tree.h
@@ -1,9 +1,9 @@
 #include <cctype>
 #include <span>
-template<size_t M> struct binary_indexed_tree {
+template<int M> struct binary_indexed_tree {
     long long b0[M], b1[M];
     int n;
-    void update(size_t l, size_t r, int x) {
+    void update(int l, int r, int x) {
         for (auto i = l; i <= n; i += i & -i) {
             b0[i] -= (long long) (l - 1) * x;
             b1[i] += x;
@@ -13,7 +13,7 @@ template<size_t M> struct binary_indexed_tree {
             b1[i] -= x;
         }
     }
-    auto query(size_t i) {
+    auto query(int i) {
         long long x = 0, y = 0;
         for (auto j = i; j; j -= j & -j) {
             x += b0[j];
@@ -21,7 +21,7 @@ template<size_t M> struct binary_indexed_tree {
         }
         return x + y * i;
     }
-    auto query(size_t l, size_t r) { return query(r) - query(l - 1); }
+    auto query(int l, int r) { return query(r) - query(l - 1); }
     void build(std::span<int> a) {
         auto n = a.size();
         this->n = n;
